refactor: コードレビュー指摘を全対応 + 品質改善
【共通ユーティリティ化】
- extractCoordinatesFromWKB を src/lib/location.ts に移行
- API RouteとUploadPageから重複コードを削除
- 修正漏れ防止

【アラート関連の改善】
- ID生成を crypto.randomUUID() に変更（衝突防止）
- タイマークリーンアップを useRef + useEffect で実装（メモリリーク防止）
- コンポーネントアンマウント時にタイマーをクリア

【アクセシビリティ向上】
- safe-area-inset で iPhoneノッチ対応
- animate-bounce に motion-reduce:animate-none を追加
- 閉じるボタンに aria-label="アラートを閉じる"

【UX改善】
- エラーメッセージを日本語化（MANHOLE_LOCATION_INVALID → ユーザー向け日本語メッセージ）
- エラーコードシステムの導入
